fix: move config to rootDir
